Split native crate into dead-cst-runtime (rlib+dylib) + cdylib shim
Move the entire native implementation into a new `dead-cst-runtime`
crate and reduce the root `dead-cst-native` crate to a thin pyo3
`#[pymodule]` shim that forwards to `dead_cst_runtime::register`.

`dead-cst-runtime` is built as both `rlib` and `dylib`:

- **rlib** — the default `_native` cdylib links this statically, so the
  shipped wheel stays self-contained and behaviorally identical (785
  tests pass unchanged).
- **dylib** — the shared-runtime artifact that external native plugins
  will link against under an opt-in `-C prefer-dynamic` build, giving
  one shared salsa/ty instance across the host and its plugins.

The dylib artifact has undefined Python symbols (pyo3 `extension-module`
skips libpython), so `.cargo/config.toml` adds `-undefined
dynamic_lookup` to its link. `prefer-dynamic` is deliberately NOT in the
committed config — the default build stays static.

No behavior change; pure restructure. Foundation for loadable external
native plugins.
