+++ Cargo.toml
@@ -16,29 +16,12 @@ crate-type = ["cdylib"]
 # / etc. at link time.
 [features]
 default = ["extension-module"]
-extension-module = ["pyo3/extension-module"]
+extension-module = ["pyo3/extension-module", "dead-cst-runtime/extension-module"]
 
 [dependencies]
 pyo3 = { version = "0.22", features = ["abi3-py311", "multiple-pymethods"] }
-salsa = "0.26"
-ty_project = { path = "vendor/ruff/crates/ty_project" }
-ty_python_core = { path = "vendor/ruff/crates/ty_python_core" }
-ty_python_semantic = { path = "vendor/ruff/crates/ty_python_semantic" }
-ty_module_resolver = { path = "vendor/ruff/crates/ty_module_resolver" }
-ruff_db = { path = "vendor/ruff/crates/ruff_db", features = ["os"] }
-ruff_python_ast = { path = "vendor/ruff/crates/ruff_python_ast" }
-ruff_python_parser = { path = "vendor/ruff/crates/ruff_python_parser" }
-ruff_source_file = { path = "vendor/ruff/crates/ruff_source_file" }
-ruff_text_size = { path = "vendor/ruff/crates/ruff_text_size" }
-ty_ide = { path = "vendor/ruff/crates/ty_ide" }
-ruff_memory_usage = { path = "vendor/ruff/crates/ruff_memory_usage" }
-get-size2 = { version = "0.8.0", features = ["smallvec"] }
-rustc-hash = "2.0"
-smallvec = { version = "1.13", features = ["union", "const_generics", "const_new"] }
-indicatif = "0.17"
-rayon = "1.10"
-crossbeam-channel = "0.5"
-regex = "1"
-serde = { version = "1", features = ["derive"] }
-bincode = "1.3"
-parking_lot = "0.12"
+# The whole implementation lives in the runtime crate. By default this is
+# linked as an rlib (static, self-contained wheel). A `-C prefer-dynamic`
+# build links it as the shared `libdead_cst_runtime.dylib` instead, which
+# external native plugins also link — giving one shared salsa/ty runtime.
+dead-cst-runtime = { path = "runtime" }
